+++ src/helpers/casts.ts
@@ -6,7 +6,12 @@ import {
 export type Balances = { [key: string]: string };
 
 /** Possible currencies types */
-export type CurrencyType = 'blockchain' | 'fiat' | 'stablecoin' | 'unknown';
+export enum CurrencyType {
+  Blockchain = 'blockchain',
+  Fiat = 'fiat',
+  Stablecoin = 'stablecoin',
+  Unknown = 'unknown',
+}
 
 /**
  * Currency type object for {@link Store.getCurrencies}
@@ -176,16 +181,18 @@ export const toCurrencies = (input: any): Currencies => {
     if (value.code) {
       const code = value.code.toString();
 
-      let type: CurrencyType = 'unknown';
-      if (value.is_blockchain) type = 'blockchain';
-      if (value.is_fiat) type = 'fiat';
-      if (value.is_stablecoin) type = 'stablecoin';
+      let type = CurrencyType.Unknown;
+      if (value.is_blockchain) type = CurrencyType.Blockchain;
+      if (value.is_fiat) type = CurrencyType.Fiat;
+      if (value.is_stablecoin) type = CurrencyType.Stablecoin;
 
       const currency: Currency = {
         name: value.name || '',
         decimals: value.decimals || 0,
         type,
       };
+
+      if (Object.prototype.hasOwnProperty.call(value, 'url')) currency.url = value.url;
 
       accumulator[code] = currency;
     }