+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - Add 'unknown' variant for `CurrencyType`
+- Replace `CurrencyType` by enum and store it in `Client` class
 - Fix library connection in examples
 - Fix links in `readme.md`
 - Fix `Client` class `setPageSize` method documentation parameter name